+++ src/commands/jira/startWorkOnIssue.ts
@@ -1,17 +1,20 @@
-import { isMinimalIssue, IssueLinkIssue, MinimalIssue, MinimalORIssueLink } from '@atlassianlabs/jira-pi-common-models';
+import { isMinimalIssue, MinimalIssue, MinimalIssueOrKeyAndSite } from '@atlassianlabs/jira-pi-common-models';
 
 import { DetailedSiteInfo } from '../../atlclients/authInfo';
 import { Container } from '../../container';
 import { fetchMinimalIssue } from '../../jira/fetchIssue';
 
-export async function startWorkOnIssue(issueOrLink: MinimalORIssueLink<DetailedSiteInfo> | undefined) {
-    let issue: MinimalIssue<DetailedSiteInfo> | undefined = undefined;
+export async function startWorkOnIssue(issueOrKeyAndSite: MinimalIssueOrKeyAndSite<DetailedSiteInfo>) {
+    let issue: MinimalIssue<DetailedSiteInfo>;
 
-    if (isMinimalIssue(issueOrLink)) {
-        issue = issueOrLink;
+    if (isMinimalIssue(issueOrKeyAndSite)) {
+        issue = issueOrKeyAndSite;
     } else {
-        const linkedIssue: IssueLinkIssue<DetailedSiteInfo> = issueOrLink as IssueLinkIssue<DetailedSiteInfo>;
-        issue = await fetchMinimalIssue(linkedIssue.key, linkedIssue.siteDetails);
+        issue = await fetchMinimalIssue(issueOrKeyAndSite.key, issueOrKeyAndSite.siteDetails);
+
+        if (!issue) {
+            throw new Error(`Jira issue ${issueOrKeyAndSite.key} not found in site ${issueOrKeyAndSite.siteDetails}`);
+        }
     }
 
     Container.startWorkWebviewFactory.createOrShow({